Fixed some Install issues and added install description
